Changed data types

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,13 +8,13 @@
 int main(void)
 {
 	int i;
-	unsigned long int fib1 = 1;
-	unsigned long int fib2 = 2;
-	unsigned long int future;
+	unsigned long long int fib1 = 1;
+	unsigned long long int fib2 = 2;
+	unsigned long long int future;
 
 	for (i = 0; i < 98; i++)
 	{
-		printf("%lu", fib1);
+		printf("%llu", fib1);
 
 		future = fib1 + fib2;
 		fib1 = fib2;
